refactor: rename Numbers Callbacks to CallbackConfiguration for clarity
